pyproject.toml
[project]
name = "analytics-assistant"
version = "0.1.0"
description = "Natural language analytics and visualization platform with MCP data source integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.54.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S"]
ignore = ["S101"]  # Allow assert statements in tests

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106"]  # Allow asserts and hardcoded passwords in tests

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
